requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-multipart>=0.0.6
pydantic>=2.0.0
Pillow>=10.0.0
httpx>=0.24.0
